[feature] reducer 설정
