+++ front/src/store/auth.js
@@ -1,0 +1,22 @@
+import { configureStore, createSlice } from '@reduxjs/toolkit'
+
+const auth = createSlice({
+    name: 'auth',
+    initialState: {
+        isAuthenticated: false,
+        user: null,
+    },
+    reducers: {
+        login: (state, action) => {
+            state.isAuthenticated = true;
+            state.user = action.payload;
+        },
+        logout: (state) => {
+            state.isAuthenticated = false;
+            state.user = null;
+        },
+    },
+});
+
+export let { login, logout } = auth.actions
+export default auth;
