+++ front/src/store/store.js
@@ -1,22 +1,12 @@
 import { configureStore, createSlice } from '@reduxjs/toolkit'
-
-
-//tablogin
-let loginTab = createSlice({
-    name: 'loginTab',
-    initialState: { name: 'customer' },
-    reducers: {
-        setLoginTab(state, action) {
-            state.name = action.payload
-        }
-    }
-})
-export let { setLoginTab } = loginTab.actions
+import loginTab from './loginTab'
+import auth from './auth'
 
 
 export default configureStore({
     reducer: {
-        loginTab: loginTab.reducer
+        loginTab: loginTab.reducer,
+        auth: auth.reducer
     }
 })
